+++ alsek/storage/backends/postgres/asyncio.py
@@ -18,9 +18,9 @@ from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, create_async_engin
 from alsek.defaults import DEFAULT_NAMESPACE
 from alsek.storage.backends import AsyncBackend, LazyClient
 from alsek.storage.backends.postgres._utils import (
-    PostgresAsyncPubSubListener,
     validate_value_within_postgres_notification_size_limit,
 )
+from alsek.storage.backends.postgres._pubsub.asyncio import PostgresAsyncPubSubListener
 from alsek.storage.backends.postgres.tables import SCHEMA_NAME, Base
 from alsek.storage.backends.postgres.tables import KeyValue as KeyValueRecord
 from alsek.storage.backends.postgres.tables import KeyValueType
